:art: Remove RPC Dir, Separate config by node type
