+++ highway/config/option.go
@@ -1,0 +1,2 @@
+package config
+
