+++ host/exchange/config.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/kataras/golog"
 	"github.com/libp2p/go-libp2p-core/protocol"
-	"github.com/sonr-io/core/config"
+	"github.com/sonr-io/core/device"
 )
 
 // Textile API definitions
@@ -65,7 +65,7 @@ func (o *options) Apply(p *ExchangeProtocol) error {
 		//mail := local.NewMail(cmd.NewClients(TextileClientURL, true, TextileMinerIdx), local.DefaultConfConfig())
 
 		// Create new mailbox
-		if config.ThirdParty.Exists(TextileMailboxDirName) {
+		if device.ThirdParty.Exists(TextileMailboxDirName) {
 			// Return Existing Mailbox
 			if err := p.loadMailbox(); err != nil {
 				return err
