+++ host/transmit/item.go
@@ -6,7 +6,7 @@ import (
 	"os"
 
 	"github.com/libp2p/go-msgio"
-	"github.com/sonr-io/core/config"
+	"github.com/sonr-io/core/motor/config"
 	v1 "go.buf.build/grpc/go/sonr-io/core/host/transmit/v1"
 	motor "go.buf.build/grpc/go/sonr-io/core/motor/v1"
 )
