+++ host/transmit/protocol.go
@@ -4,8 +4,9 @@ import (
 	"context"
 
 	"github.com/libp2p/go-libp2p-core/network"
-	"github.com/sonr-io/core/config"
+	"github.com/sonr-io/core/device"
 	"github.com/sonr-io/core/host"
+	"github.com/sonr-io/core/motor/config"
 	"github.com/sonr-io/core/util"
 	v1 "go.buf.build/grpc/go/sonr-io/core/host/transmit/v1"
 	motor "go.buf.build/grpc/go/sonr-io/core/motor/v1"
@@ -19,7 +20,7 @@ type TransmitProtocol struct {
 	node     host.HostImpl
 	ctx      context.Context // Context
 	current  *v1.Session     // current session
-	mode     config.Role
+	mode     device.Role
 }
 
 // New creates a new TransferProtocol
@@ -85,7 +86,7 @@ func (p *TransmitProtocol) Outgoing(payload *types.Payload, to *types.Peer) erro
 	p.current = NewOutSession(payload, from, to)
 
 	// Send Files
-	if util.IsFile(p.current.Payload.GetItems()[0].GetMime()) {
+	if p.current.Payload.GetItems()[0].GetMime().Type != types.MIME_TYPE_URL {
 		// Create New Stream
 		stream, err := p.node.NewStream(p.ctx, toId, FilePID)
 		if err != nil {
