--- util/errors.go
@@ -1,8 +1,0 @@
-package util
-
-import "errors"
-
-var (
-	ErrFragmentAndQuery = errors.New("Unable to create new DID. Fragment and Query are mutually exclusive")
-	ErrParseInvalid     = errors.New("Unable to parse string into DID, invalid format.")
-)
